make sure dependencies and tests work with Ruby 2.1.4 (v0.5.0 prep)
We are aiming to support Chef `12.1+`, meaning we need to support Ruby
`2.1.4+` within this cookbook. This includes updating our dependencies
ad our `.travis.yml` file.

This upgrades our Gemfile so that we're locked to versions supporting
2.1.4 and Chef 12.1.0+:

* chef `12.1.0` if none is provided in the environment
* foodcritic `~> 7.1`, as `8.0+` isn't compatible
* rack `~> 1.6`,  as `2.0+` isn't compatible
* berkshelf `~> 4.3`, as `5.0` isn't compatible

In addition to the changes made above, I've updated the `metadata.rb`
file to include two further pieces of metadata to appease foodcritic:
`issues_url` and `source_url`.

Lastly, the `spec_helper` file for the ServerSpec tests has been updated
to be ServerSpec2 compatible.